requirements.txt
anthropic>=0.25.0
feedparser>=6.0.11
notion-client>=2.2.1
python-dotenv>=1.0.0
python-dateutil>=2.9.0
requests>=2.31.0
